GraphQL Updates
- Breaking apart GraphQL resolvers into multiple files
- Co-Locate schema with resolvers
- Examples of capturing GraphQL requests and responses
- Use `schema.json` for `.eslintrc.js` and `apollo.config.js`
