+++ apollo.config.js
@@ -1,14 +1,8 @@
-const fs = require('fs')
-const tempfile = require('tempfile')
-const { genSchema } = require('./src/main/webapp/intrigue-api/gen-schema')
-const path = tempfile('schema.graphql')
-fs.writeFileSync(path, genSchema())
-
 module.exports = {
   client: {
     service: {
       name: 'intrigue-api',
-      localSchemaFile: path,
+      localSchemaFile: './schema.json',
     },
     includes: ['src/**/*.{ts,tsx,js,jsx}'],
   },
