+++ package.json
@@ -15,6 +15,7 @@
     "eslint-plugin-graphql": "^3.1.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-react-hooks": "^2.1.2",
+    "json-diff": "^0.5.4",
     "mocha": "^5.2.0",
     "node-fetch": "^2.6.0",
     "tempfile": "^3.0.0",
