+++ src/main/webapp/intrigue-api/metacards/cql.js
