+++ src/main/webapp/utils/index.js
@@ -1,1 +1,1 @@
-export { mergeDeep, mergeDeepOverwriteLists } from './immutable-utils'
+module.exports = require('./immutable-utils')
